feat: Move addForceEncryptedPlaceHolder to fields helper
